requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
cffi==1.17.1
Django==5.1.4
django-environ==0.11.2
django-extensions==3.2.3
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pycparser==2.22
sqlparse==0.5.3
tzdata==2024.2
websockets==14.1
Werkzeug==3.1.3
